Renovation: remove testing/jest folder, test utils refactoring (#13910)
* Renovation: remove testing/jest folder, test utils refactoring

Revert "Renovation: remove testing/jest folder, test utils refactoring"

This reverts commit 0300381e3de312c340beabaffeb3178669219e3b.

fix

* refactoring

* fix generation task

* rebase
