+++ jest.config.js
@@ -25,19 +25,18 @@ module.exports = {
             branches: 0
         }
     },
-    roots: ['<rootDir>/testing/jest', '<rootDir>/js/renovation'],
+    roots: ['<rootDir>/js/renovation'],
     moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
     preset: 'ts-jest',
     setupFiles: [
-        path.join(path.resolve('.'), './js/renovation/__tests__/setup_enzyme.ts'),
+        path.join(path.resolve('.'), './js/renovation/test_utils/setup_enzyme.ts'),
     ],
     testMatch: [
-        '<rootDir>/testing/jest/**/*.tests.[jt]s?(x)',
         '<rootDir>/js/renovation/**/__tests__/**/*.test.[jt]s?(x)'
     ],
     transform: {
-        'test_components.+\\.tsx$': path.resolve('./js/renovation/__tests__/transformers/declaration.js'),
+        'test_components.+\\.tsx$': path.resolve('./js/renovation/test_utils/transformers/declaration.js'),
         '\\.(js|jsx|ts)$': resolve.sync('ts-jest'),
-        '\\.(tsx)$': path.resolve('./js/renovation/__tests__/transformers/tsx.js')
+        '\\.(tsx)$': path.resolve('./js/renovation/test_utils/transformers/tsx.js')
     }
 };
