+++ js/renovation/preact_wrapper/__tests__/utils/test_components/preact_test_widget.tsx
@@ -11,7 +11,7 @@ import {
 } from 'devextreme-generator/component_declaration/common';
 import {
   keyboard,
-} from '../../../../events/short';
+} from '../../../../../events/short';
 
 export const view = (viewModel: PreactTestWidget) => (
   <div
