+++ js/renovation/test_utils/events_mock.ts
