+++ js/renovation/test_utils/transformers/declaration.js
