+++ js/renovation/ui/common/__tests__/widget.test.tsx
@@ -4,7 +4,7 @@ import { shallow } from 'enzyme';
 import {
   clear as clearEventHandlers, defaultEvent, emit,
   emitKeyboard, getEventHandlers, EVENT, KEY,
-} from '../../../__tests__/events_mock';
+} from '../../../test_utils/events_mock';
 import { Widget, viewFunction, WidgetProps } from '../widget';
 import { isFakeClickEvent } from '../../../../events/utils';
 import config from '../../../../core/config';
